+++ hdr/Map.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "Structs/SharedContent.h"
+
+void InitMap(SharedContent* Content);
+void FreeMap(SharedContent* Content);
