+++ hdr/Overwrite.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <stdio.h>
+#include "Structs/IntTuple.h"
+
+void Overwrite(IntTuple pos, IntTuple endl, char z);
