+++ hdr/Structs/DataS.h
@@ -1,0 +1,10 @@
+#pragma once
+#include <windows.h>
+#include <stdbool.h>
+#include "Structs/SharedContent.h"
+
+typedef struct DataS
+{
+    SharedContent* content;
+    bool* GameOver;
+} DataS;
