+++ hdr/Structs/FullData.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "Structs/DataS.h"
+#include "Structs/Queue.h"
+
+typedef struct FullData
+{
+    SharedContent* content;
+    bool* GameOver;
+    Queue** Tail;
+} FullData;
