+++ hdr/Structs/IntTuple.h
@@ -1,0 +1,7 @@
+#pragma once
+
+typedef struct IntTuple
+{
+    int x;
+    int y;
+}IntTuple;
