+++ hdr/Structs/QueueBase.h
@@ -1,0 +1,8 @@
+#pragma once
+
+typedef struct QueueBase
+{
+    int xpos;
+    int ypos;
+    struct QueueBase* Next;
+} QueueBase;
