+++ hdr/Tail.h
@@ -1,9 +1,7 @@
 #pragma once
-
-#include "Shared.h"
-
-typedef struct Queue Queue;
-typedef struct IntTuple IntTuple;
+#include "Structs/SharedContent.h"
+#include "Structs/Queue.h"
+#include "Overwrite.h"
 
 void RotateTail(SharedContent* content, Queue** Tail);
 void AddTailSegment(SharedContent* content, Queue** Tail);
