+++ hdr/Threads.h
@@ -1,19 +1,5 @@
-#include <windows.h>
-#include "Snake.h"
-#include "Shared.h"
-
-typedef struct DataS
-{
-    SharedContent* content;
-    bool* GameOver;
-} DataS;
-
-typedef struct FullData
-{
-    SharedContent* content;
-    bool* GameOver;
-    Queue** Tail;
-} FullData;
+#pragma once
+#include "Structs/FullData.h"
 
 DWORD WINAPI MovementThread(LPVOID pData);
 DWORD WINAPI DirectionThread(LPVOID pData);
